refactor skill configs to using skill values (#127)

+++ Assets/Main/Scripts/Skills/Common/Component/Config/Value/SkillValue.cs
@@ -1,4 +1,3 @@
-using Fusion;
 using Main.Scripts.Skills.Common.Component.Config.Value.Resolver;
 using UnityEngine;
 
@@ -6,7 +5,7 @@ namespace Main.Scripts.Skills.Common.Component.Config.Value
 {
 public abstract class SkillValue : ScriptableObject
 {
-    public int Resolve(SkillVariableProvider variableProvider, GameObject? target = null)
+    public float Resolve(SkillVariableProvider variableProvider, GameObject? target = null)
     {
         return SkillValueResolver.Resolve(this, variableProvider, target);
     }
